federation-api: Use SinglePath as PathBuilder of requests
