+++ crates/ruma-federation-api/src/backfill/get_backfill.rs
@@ -18,9 +18,7 @@ pub mod v1 {
         method: GET,
         rate_limited: false,
         authentication: ServerSignatures,
-        history: {
-            1.0 => "/_matrix/federation/v1/backfill/{room_id}",
-        }
+        path: "/_matrix/federation/v1/backfill/{room_id}",
     }
 
     /// Request type for the `get_backfill` endpoint.
