+++ crates/ruma-federation-api/src/event/get_room_state_ids.rs
@@ -16,9 +16,7 @@ pub mod v1 {
         method: GET,
         rate_limited: false,
         authentication: ServerSignatures,
-        history: {
-            1.0 => "/_matrix/federation/v1/state_ids/{room_id}",
-        }
+        path: "/_matrix/federation/v1/state_ids/{room_id}",
     }
 
     /// Request type for the `get_room_state_ids` endpoint.
