+++ crates/ruma-federation-api/src/keys/claim_keys.rs
@@ -21,9 +21,7 @@ pub mod v1 {
         method: POST,
         rate_limited: false,
         authentication: ServerSignatures,
-        history: {
-            1.0 => "/_matrix/federation/v1/user/keys/claim",
-        }
+        path: "/_matrix/federation/v1/user/keys/claim",
     }
 
     /// Request type for the `claim_keys` endpoint.
