+++ crates/ruma-federation-api/src/keys/get_keys.rs
@@ -21,9 +21,7 @@ pub mod v1 {
         method: POST,
         rate_limited: false,
         authentication: ServerSignatures,
-        history: {
-            1.0 => "/_matrix/federation/v1/user/keys/query",
-        }
+        path: "/_matrix/federation/v1/user/keys/query",
     }
 
     /// Request type for the `get_keys` endpoint.
