+++ crates/ruma-federation-api/src/membership/create_join_event/v2.rs
@@ -13,9 +13,7 @@ metadata! {
     method: PUT,
     rate_limited: false,
     authentication: ServerSignatures,
-    history: {
-        1.0 => "/_matrix/federation/v2/send_join/{room_id}/{event_id}",
-    }
+    path: "/_matrix/federation/v2/send_join/{room_id}/{event_id}",
 }
 
 /// Request type for the `create_join_event` endpoint.
