+++ crates/ruma-federation-api/src/membership/create_leave_event/v2.rs
@@ -12,9 +12,7 @@ metadata! {
     method: PUT,
     rate_limited: false,
     authentication: ServerSignatures,
-    history: {
-        1.0 => "/_matrix/federation/v2/send_leave/{room_id}/{event_id}",
-    }
+    path: "/_matrix/federation/v2/send_leave/{room_id}/{event_id}",
 }
 
 /// Request type for the `create_leave_event` endpoint.
