+++ crates/ruma-federation-api/src/membership/prepare_leave_event.rs
@@ -18,9 +18,7 @@ pub mod v1 {
         method: GET,
         rate_limited: false,
         authentication: ServerSignatures,
-        history: {
-            1.0 => "/_matrix/federation/v1/make_leave/{room_id}/{user_id}",
-        }
+        path: "/_matrix/federation/v1/make_leave/{room_id}/{user_id}",
     }
 
     /// Request type for the `get_leave_event` endpoint.
