+++ crates/ruma-federation-api/src/query/get_room_information.rs
@@ -16,9 +16,7 @@ pub mod v1 {
         method: GET,
         rate_limited: false,
         authentication: ServerSignatures,
-        history: {
-            1.0 => "/_matrix/federation/v1/query/directory",
-        }
+        path: "/_matrix/federation/v1/query/directory",
     }
 
     /// Request type for the `get_room_information` endpoint.
