+++ crates/ruma-federation-api/src/transactions/send_transaction_message.rs
@@ -23,9 +23,7 @@ pub mod v1 {
         method: PUT,
         rate_limited: false,
         authentication: ServerSignatures,
-        history: {
-            1.0 => "/_matrix/federation/v1/send/{transaction_id}",
-        }
+        path: "/_matrix/federation/v1/send/{transaction_id}",
     }
 
     /// Request type for the `send_transaction_message` endpoint.
